requirements.txt
joblib==1.2.0
numpy==1.23.5
opencv-python-headless
pandas==1.5.3
Pillow==9.5.0
scikit-image
streamlit==1.22.0
tensorflow
scikit-learn